net/mlx4_core: Memory allocation based on NUMA
This is a fix for current allocation method, which does not take into
account the index distribution of cores to NUMA.
Now, we use the affinity configuration from the previous commit, set for
each ring to identify it NUMA (called ring NUMA).

We pay attention to the following allocations to be based on NUMA:

1. TX/RX CQ/ring SW struct.
2. HW buffers - CQ buffer, Send/Recv WQE buffer, Blue Flame buffer, etc..
3. Data Buffers allocated for the incoming packets.

1,3 and CQ buffer will be allocated on ring NUMA (SW read), while WQE
buffer, UARs, and BF bufffers allocated on device NUMA (HW read).

Ported from 3.4, originates at 2.4.

issue: 349720
Change-Id: I60fcba7e4b27a3c609744f921b51a72b8a0f03bc
Signed-off-by: Ido Shamay <idos@mellanox.com>
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>