IB/mlx4: QP should register to mcast for all ports if its port is unknown
When QP is in RESET/INIT state, its port isn't set (in RESET) or is set
but could be further changed.
In addition, when registering to mcast and working in DMFS,
a flow steering rule is created.
Creating such rule require the port, theerefore we need to
register the rule on all ports.
Previously, the driver registered with port = 0, which is errnous
behaviour.

This breaks clause 11.3.1 ATTACH QP TO MULTICAST GROUP of the IB
spec which requries to only attach the required HCA port. Since we're
already breaking this behaviour for QPs in state RESET and INIT (ARCH decision),
we simplify the implementation by using the same behaviour for all
QP states.

This patch is NOT FOR UPSTREAM!

issue: 327456
issue: 874811
Change-Id: I33e6ded8660938ebc95140bca480976a17348beb
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>