IB/mlx4: Add ioctl support
Add support for get hw clock params through ioctl

ioctls are provider specific and routed straight to the provider driver.
This was done in order to support getting the hardware
clock offset from user space.

Note: when libmlx4 will be changed to get this information via uhw this
patch can be dropped.

issue: 650074
Change-Id: I08be7a948d73c65456478e032690537ea0ef950a
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>