IB: Add QP attribute 'flow_entropy'
The QP attribute 'flow_entropy' can be used to help the application
to put QPs into different classes for different purposes.
In RoCE LAG we use this value to put all QPs with the same
flow_entropy on the same port (instead of load balancing it).
This is useful when the application has knowledge about which port is best for a QP to
achieve equal balancing or low latency for example.

This is based on Moni's commit from 3.1 (4361f7cf)

issue: 654966
Change-Id: I71ccd70ea09d4ed70d072096fdb70e834934213f
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>