net/mlx4_core: Notifying user when using an obsolete module patameter
The 'high_rate_steer' parameter was present in mlnx_ofed_2_* releases,
it is now covered by 'Enable optimized steering' option in
log_num_mgm_entry_size module parameter.
If this param is used, a message will notify the user that the
parameter is obsolete and mlx4_core will not load. Previously,
mlx4_core was not loaded without any notification to the user.

issue: 657967
Change-Id: I3a91549dbb30df06c043fe406bceeee5771c9e70
Signed-off-by: Noa Osherovich <noaos@mellanox.com>
Signed-off-by: Daniel Jurgens <danielj@mellanox.com>