IB/core: Add mic node and scif transport types
The OFED SCIF driver implements a software-emulated RDMA device to
allow OFED based applications, such as Intel MPI, to run on Intel(R)
MIC Architecture without the presence of a physical HCA.
OFED SCIF is only targeted for inter- node communication within a
single platform, where a node is a coprocessor or the host processor.
This patch adds new node and transport types to the ib_core kernel
module to distinguish this new RDMA interface type.
This code was received from Intel and will be introduced in a series of
patches.

This patch also introduces two new function which help with testing for
scif capability.

Issue: 560974
Issue: 890553
Change-Id: Ibae761541951b1e440907de1924549f5b7747c5f
Signed-off-by: Noa Osherovich <noaos@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Mark Bloch <markb@mellanox.com>