IB/mlx4: Assign user allocated DB and BF to QP
Rebase 4.0, based on Moshe's code from 3.x.

Motivation: assigning dedicated DB and BF to a QP saves usage of
barriers and locks on data path.

On create QP, the user pass the virtual address of the pre allocated
DB UAR.
The driver checks that the virtual address is related to one of the
user UARs.
If so, the driver assign the UAR to the new QP.

issue: 862323
Change-Id: Ic62be1e6053f6b59576450f27937034d059a303d
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>