IB/core: Introduce peer client interface
Rebase:
Based on V2 that was sent upstream.

Introduces an API between IB core to peer memory clients,(e.g. GPU
cards) to provide access for the HCA to read/write GPU memory.

As a result it allows RDMA-based application to use GPU computing power,
and RDMA interconnect at the same time w/o copying the data between the
P2P devices.

Each peer memory client should register with IB core. In the
registration request, it should supply callbacks to its memory basic
functionality such as get/put pages,get_page_size, dma map/unmap.

The client can optionally require the ability to invalidate memory it
provided, by requesting an invalidation callback details.

Upon successful registration, IB core will provide the client with a
unique registration handle and an invalidate callback function in case required
by the peer.

The handle should be used when unregistering the client, the callback
function can be used by the client in later patches, for a request from the
client to immediately release pinned pages.

Each peer must be able to recognize whether it's the owner of
a specific virtual address range. In case the answer is YES, further
calls for memory functionality will be tunneled to that peer.

The recognition is done via the 'acquire' call. The call arguments
provide the address and size of the memory requested. In case peer-direct context
information is available from the user verbs context, it is provided as well.
Upon recognition, the acquire call returns a peer direct client specific
context.

The context will be provided by the peer direct controller to the peer
direct client callbacks when referring the specific address range.

issue: 862323
Change-Id: I920c4a878af21a5241f9e2f377f9a11d07785de6
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>