IB/core: Sysfs support for peer memory
Supplies the required functionality to expose information and
statistics over sysfs for a given peer memory client.

This mechanism enables userspace application to check
which peers are available (based on name & version) and based on that
decides whether it can run successfully.

Root sysfs directory is /sys/kernel/mm/<peer_name>, under that
directory will reside some files that represent the statistics for that peer.

issue: 862323
Change-Id: I19f19b3b84f8ac5551deb6e1678309f055f432b8
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>