net/mlx4_core: Add counters mode to slave meta-data
The counters mode is global for all the counters in the device. Add
Infrastructure to support extended counters mode.

To allow compatibility when old VF is running over a new hypervisor
configured to work in extended counters mode, the PF should hold the
counters mode (basic/extended) for every slave. So it will be able to
translate counter data between modes if the VF expects basic counter.

Issue: 418563
Issue: 867998
Change-Id: I89f923fb757ef71856dee25e4c49f7b233b6a974
Signed-off-by: Eran Ben Elisha <eranbe@mellanox.com>
Signed-off-by: Hadar Hen Zion <hadarh@mellanox.com>
Signed-off-by: Tariq Toukan <tariqt@mellanox.com>