IB/mlx4: Add support for RSS QP from non-experimental flow
After the changes during the rebase it is no longer possible to
create a QPG from ib_create_qp. Added a simple solution to
to create an RSS parent QP and pass an experimental init_attr.
This will allow us to support RSS without duplicating code with
minimal impact on the base code.

issue: 915726

Change-Id: Id9562a6de46d915f7413fe87219f6abd6d8a1854
Signed-off-by: Alex Vesker <valex@mellanox.com>