IB/ipoib: Set device address using sysfs instead of ndo
In order to support IB router, a user must be able to
set the device address of an IPoIB interface. such operation
is usually done via set_mac_address ndo.

There is a bug in the bonding driver that is triggered when
an IPoIB exposes this ndo. This patch is a workaround that fixes
the issue by using sysfs instead of exposing the ndo function.

Usage Example:
  echo 11:11:11:11:11:11:11:11 > /sys/class/net/ib0/set_mac

issue: 814941
issue: 932967

Change-Id: I81c2bce8e92a05a6bb2e4c33d0ce534a467a01de
Signed-off-by: Feras Daoud <ferasda@mellanox.com>
Signed-off-by: Mark Bloch <markb@mellanox.com>