IB/mlx5: Set rx_hash_symmetric according to spec
RSS is defined in MSDN spec to use a non symmetric hash. Mellanox's
internal implementation uses a symmetric hash by default which is not
according to spec and doesn't have an API to change it. As a result,
customers who calculate the hash according to the spec and compare it
to the result exposed by DPDK get different values.

Change the default value of rx_hash_symmetric from 1 to 0 so that
until we have an API to change it, it will act according to the spec
and to customers' expectations.

issue: 941069
Change-Id: I4d65f374a021a910fba4bb0a0e8237634a049ada
Signed-off-by: Noa Osherovich <noaos@mellanox.com>