net/mlx4_core: Don't fail fw SET_IF_STAT command when running over Windows Hypervisor
The upstream kernel (and Mellanox OFED 4.x) use a handshake mechanism
with guests for SET_IF_STAT. This mechanism is not implemented in
the Windows Hypervisor (which returns a BAD_PERM fw error code
when a guest attempts to invoke the command).

The Windows Hypervisor sets the counters mode to extended.

Recognize the -EPERM return (in addition to the -EINVAL return)
on the guest as indicating that it is running over an old
Hypervisor which sets counters to extended mode.

issue: 951840
Change-Id: I23a56ad5c4bf98010b551debc65886465c379a67
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>