IB/ipoib: Fix race condition in neigh creation
When using RSS/TSS, two threads may execute xmit in parallel
to two different TX queues while the target is the same. In this
case, both of them will add the same neighbor to the path's neigh
link list and will create an endless loop.

Analisis
Two SKB are scheduled to be transmitted from two cores.
In ipoib_start_xmit, both gets NULL when calling ipoib_neigh_get.
Two calls to neigh_add_path are made. One thread takes the spin-lock
and calls ipoib_neigh_alloc which creates the neigh structure,
then (after the __path_find) the neigh is added to the path's neigh
link list. When the second thread enters the critical section it also
calls ipoib_neigh_alloc but in this case it gets the already allocated
ipoib_neigh structure, which is already linked to the path's neigh
link list and adds it again to the list. Which beside of triggering
the list, it creates a loop in the linked list. This loop leads to endless
loop inside path_rec_completion.

Solution:
Check list_empty(&neigh->list) before adding to the list.
Add a similar fix in "ipoib_multicast.c::ipoib_mcast_send"

issue: 165686

Change-Id: If50f44beaa898f75e465c534a07ad7fac09a9106
Signed-off-by: Feras Daoud <ferasda@mellanox.com>
Signed-off-by: Shlomo Pongratz <shlomop@mellanox.com>