IB/ipoib: Force CQ drain when destroying tx
When the send-queue is closed and there are cqe's that are missing, (it
can be after race between closing the queue and moving the qp
to error state (after RNR_EXC event) by the FW).
In that situation the napi will not run and the CQ has nothing for us so
arming it will not help, the only way to take the driver out of that
situation is by draining the cq.
The TX destroy function will verify that there are no cqe's that are in
the CQ.
In that case make sure all the wqe's are handled by draining the CQ.

issue:946836
Change-Id: I408e6eec1b560ffe4cfdf1d80b268956c1b6504f
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>
Signed-off-by: Noa Osherovich <noaos@mellanox.com>