net/mlx5_core: Add icmd support
mlx5 supported devices can be querried and controlled through configuration
cycles. This can be utilized to access and control the device before the
driver loads and there is a set of supported commands that can be issued.

This patch implements a generic icmd interface that can be used by the
driver.

Since only CAPI uses icmd, only enable with IBM kernel (based on
CONFIG_CXL_LIB config flag)

Issue: 978729
Change-Id: Iaef60ab66f00560d3fd6021091b3fcf61a2eafc6
Signed-off-by: Eli Cohen <eli@mellanox.com>