net/mlx5: Disable vfs pages counter warning on internal error
When there is internal error we flush the rb pages tree manually.
In PPC, VFs and PF may share the same 64K page and therefore the vfs
pages counter may be wrong after the reclaimng of the pages.
The fw pages counter will still be valid.
The VFs pages counter warning will be disabled on internal error.

Issue: 959407
Change-Id: I443748f84740c0b11defdef8de93d77d15cb256c
Signed-off-by: Mohamad Haj Yahia <mohamad@mellanox.com>