net/mlx4_core: Fix wrong assumption of flow steering rules type
Currently, we assume every flow steering rule is of NORMAL type and we
check the accompanied flow steering specification.
This is wrong as both the PF and VF could create any rule they want.
Fix that by denying VF of creating rules other than NORMAL and
MC_DEFAULT, and skip the specification check if the given rule is not a
NORMAL rule.

issue: 538563
issue: 875551
Change-Id: Ia056a98f89d0a0b550f0ca7ee307be056f7bdc10
Signed-off-by: Matan Barak <matanb@mellanox.com>
Signed-off-by: Gal Pressman <galp@mellanox.com>