net/mlx5: FPGA, Add client driver interface
mlx5_core handles interaction with the FPGA shell hardware, but not
with the SBU.

Add interface for Innova FPGA client drivers to interact with the SBU.

A client driver registers to the FPGA core, gets notifications of devices
in the system, and may call any of the in-kernel "SDK" functions to
interact with its hardware logic.

This allows further extending netdevice capabilities, accelerations, and
utilizing other Innova FPGA logic.

Issue: 1005200

Change-Id: Ida5364be92c69f09217e4013da378281a7e909bf
Signed-off-by: Ilan Tayari <ilant@mellanox.com>
Signed-off-by: Boris Pismenny <borisp@mellanox.com>
Signed-off-by: Yevgeny Kliteynik <kliteyn@mellanox.com>