IB/ipoib: Merge INLINE_SIZE and INLINE_THOLD
WQE used to have to parameter of inline,
one that limit the WQE size on creation and
the other used as threshold to decide whether using
the inline or not.

As a result the inline was used only for message
data under 120 while the size of the inline was 800,
causing problems that was mentioned in the attached issue.

Both sized changed to the same size 120 and merged.

issue: 886431

Change-Id: I33ada9883fe751f1d90040db45e94fa2e828b713
Signed-off-by: Shalom Lagziel <shaloml@mellanox.com>