net/mlx5e: Align ratelimit feature with upstream code
When ratelimit feature was introduced, a necessary if statement was
removed in switch_priv_channels and resulted in a bug and diverged base
code.

This patch will bring back the if statement that makes sure new skbs
will not be assigned to queues that are about to be destroyed in
select_queue.
Also, make ratelimit cleanup after itself on init failure.

Issue: 1042541
Change-Id: If40fa3058c59d15b14b4a38c3ca03f2a34277c37
Signed-off-by: Gal Pressman <galp@mellanox.com>