mlx5: CAPI implementation
Implement the initalization and teardown flows of capi.

In capi we have a specific function which is designated as the capi owner.
Each function will attempt to become the owner but firmware will choose on
of the functions and report that through the access register command.

The initialization of CAPI is done prior to driver initalization and is
using access register commands to access the CAPI desginated register to
execute the initialization flow.

Currently, swtiching the CAPP to PCI mode fails. In this case avoid
closing the CAPI BAR as the CAPP will continue to send invalidations and
those must be answered.

Issue: 978729
Change-Id: If752ae0c57560d948297225734ebbf02bbae5468
Signed-off-by: Eli Cohen <eli@mellanox.com>