net/mlx4_en: fix wrong bounderies check in ethtool set channels
When request to change the number of channels, need to check the same
values which were set in mlx4_en_get_channels.

The values that were tested before this patch, were outdated and did not
hold the correct values. This fix will maintain a better
robustness as max_tx/tx will be held in one location.

Issue:1148780
Change-Id: I73d19204f28497453baaa0f261bcf965a1b557bd
Signed-off-by: Eran Ben Elisha <eranbe@mellanox.com>