IB/uverbs: Add device memory mr registration via exp_reg_mr
This change add support to register allocated device memory
buffer (DM) as MR to be accessed locally or remotely.

The uapi command of exp_reg_mr is extended to include a
new parameter - dm_handle - which is used to point the kernel
to the relevant ib_dm object for the registration.

The ib_mr_init_attr struct, used in the call to the provider's
handler of exp_reg_user_mr is extended to include a pointer to
ib_dm object and a new mr type is added - IB_MR_TYPE_DM.

Issue: 876310
Change-Id: Ic319226b58898a9f0feaa1067aebdb5ff17f8cc1
Signed-off-by: Ariel Levkovich <lariel@mellanox.com>