IB/ipoib: RSS/TSS, Get rid of the tx_outstanding variable in all modes
The first step toward using NAPI in the UD/TX flow is to separate
between two flows, the NAPI and the xmit, meaning no use of shared
variables between both flows.

This patch takes out the tx_outstanding variable that was used in both
flows and instead the driver uses the 2 cyclic ring variables: tx_head
and tx_tail, tx_head used in the xmit flow and tx_tail in the NAPI flow.

Issue:1086647
Change-Id: Iaf24947de9da86fa525ef9b785f9c3be0b2a8479
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>