net/mlx4_core: Support IB ports for VMs with old drivers
In some older VM drivers, the counter-index field remains
uninitialized in the IB driver INIT-to-RTR qp transition. This causes
the transition to fail, and the VM driver does not come up.
Detect this case (VM submits zero counter index) in the Hypervisor
and substitute the sink counter index for the zero index.

issue: 1178129
Change-Id: I51785d6a3fbbb9bcbacdfb5761594c6bd08d8cb6
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>